renamed to ghga_event_schemas (GDEV-1319) (#19)
to accommodate the transition to Apache Kafka.
Moreover, updated dev dependencies and static files and prepared for the deprecation of dev branch.

Bump version to 0.4.0.
