+++ .devcontainer/docker-compose.yml
@@ -11,7 +11,7 @@ services:
         # [Choice] Install Node.js
         INSTALL_NODE: "true"
         NODE_VERSION: "lts/*"
-        PACKAGE_NAME: "ghga_message_schemas"
+        PACKAGE_NAME: "ghga_event_schemas"
         # On Linux, you may need to update USER_UID and USER_GID below if not your local UID is not 1000.
         USER_UID: 1000
         USER_GID: 1000
