+++ .flake8
@@ -16,5 +16,5 @@
 [flake8]
 ignore = E, W # ignore all style checks from pycodestyle
               # as they are already checked by black
-exclude = .git,__pycache__,db_migration,build,dist
+exclude = .git,__pycache__,build,dist
 max-complexity = 10
