+++ .github/workflows/check_mandatory_and_static_files.yaml
@@ -6,9 +6,9 @@ jobs:
   check-mandatory-and-static-files:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v1
+      - uses: actions/checkout@v2.4.2
       - name: Set up Python 3.9
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v3.1.2
         with:
           python-version: 3.9
       - name: Check mandatory and static files
