+++ .github/workflows/unit_and_int_tests.yaml
@@ -8,10 +8,10 @@ jobs:
     name: Unit and Integration Tests
 
     steps:
-      - uses: actions/checkout@v1
+      - uses: actions/checkout@v2.4.2
 
       - id: common
-        uses: ghga-de/gh-action-common@v1
+        uses: ghga-de/gh-action-common@v1.0.2
 
       - name: Run pytest
         run: |
@@ -20,8 +20,10 @@ jobs:
             --cov="${{ steps.common.outputs.MAIN_SRC_DIR }}" \
             --cov-report=xml
 
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v2
-        with:
-          fail_ci_if_error: true
-          verbose: true
+      - id: coveralls
+        name: Upload coverage to coveralls
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          pip3 install --upgrade coveralls
+          coveralls --service=github
