+++ .gitignore
@@ -128,5 +128,10 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
-# ingore vscode settings:
+# ignore VS Code settings:
 .vscode/
+
+# desktop settings and thumbnails
+.DS_Store
+desktop.ini
+thumbs.db
