+++ .mypy.ini
@@ -1,0 +1,12 @@
+# Global options:
+
+[mypy]
+disable_error_code = import
+show_error_codes = True
+exclude = (?x)(
+    build/lib/
+  )
+warn_redundant_casts = True
+warn_unused_ignores = True
+check_untyped_defs = True
+no_site_packages = False
