+++ .pre-commit-config.yaml
@@ -37,32 +37,32 @@ repos:
       - id: debug-statements
       - id: debug-statements
   - repo: https://github.com/psf/black
-    rev: 21.10b0
+    rev: 22.3.0
     hooks:
       - id: black
   - repo: https://github.com/pycqa/isort
-    rev: 5.10.0
+    rev: 5.10.1
     hooks:
       - id: isort
-        args: ["--profile", "black"]
+        args: [--profile, black]
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v0.910-1
+    rev: v0.981
     hooks:
       - id: mypy
-        args: [--show-error-codes, --disable-error-code, import]
+        args: [--no-warn-unused-ignores]
   - repo: https://github.com/PyCQA/pylint
-    rev: "v2.11.1"
+    rev: v2.13.3
     hooks:
       - id: pylint
         args: [--disable=E0401]
-        exclude: db_migration|tests|.devcontainer
+        exclude: tests|.devcontainer
   - repo: https://github.com/PyCQA/flake8
-    rev: "4.0.1"
+    rev: 4.0.1
     hooks:
       - id: flake8
         args: [--config, .flake8]
   - repo: https://github.com/PyCQA/bandit
-    rev: "1.7.0"
+    rev: 1.7.4
     hooks:
       - id: bandit
         exclude: tests|.devcontainer|scripts
