+++ .static_files
@@ -17,7 +17,6 @@ scripts/get_package_name.py
 scripts/check_mandatory_and_static_files.py
 scripts/README.md
 
-.github/workflows/check_pr_target.yaml
 .github/workflows/static_code_analysis.yaml
 .github/workflows/check_mandatory_and_static_files.yaml
 .github/workflows/unit_and_int_tests.yaml
@@ -30,7 +29,7 @@ example_data/README.md
 .gitignore
 .pre-commit-config.yaml
 .pylintrc
-codecov.yml
+.mypy.ini
 
 LICENSE
 requirements.txt
