--- codecov.yml
@@ -1,11 +1,0 @@
-coverage:
-  status:
-    project:
-      default:
-        target: 80% # the required coverage value
-        threshold: 1% # the leniency in hitting the target
-
-comment:
-  behavior: new
-  require_head: no # default yes
-  require_base: no # default yes
