+++ ghga_event_schemas/json_schemas/drs_object_registered.json
@@ -1,12 +1,16 @@
 {
-  "$id": "https://raw.githubusercontent.com/ghga-de/milestones-docs/main/milestone_1/api_definitions/async_topics/file_internally_registered.json",
+  "$id": "https://raw.githubusercontent.com/ghga-de/milestones-docs/main/milestone_1/api_definitions/async_topics/drs_object_registered.json",
   "$schema": "https://json-schema.org/draft/2020-12/schema",
   "additionalProperties": false,
-  "description": "This event type is triggered when a new file is registered to the Internal File Registry service and is available in the Permanent Storage.",
+  "description": "This event type is triggered when a new DRS object is registered to the DRS3 service.",
   "properties": {
     "creation_date": {
       "description": "Timestamp (in ISO 8601 format) when the entity was created.",
       "format": "date-time",
+      "type": "string"
+    },
+    "drs_uri": {
+      "description": "The DRS URI of pattern `drs://<drs_service_hostname>/<drs_id>`.",
       "type": "string"
     },
     "file_id": {
@@ -15,10 +19,6 @@
     },
     "format": {
       "description": "The format of the file: BAM, SAM, CRAM, BAI, etc.",
-      "type": "string"
-    },
-    "grouping_label": {
-      "description": "This is a label that might be use to group multiple files together. E.g. the study id.",
       "type": "string"
     },
     "md5_checksum": {
@@ -30,7 +30,7 @@
       "type": "integer"
     },
     "timestamp": {
-      "description": "The time when the message was published.",
+      "description": "The time when the event was published.",
       "format": "date-time",
       "type": "string"
     },
@@ -42,14 +42,14 @@
   },
   "required": [
     "creation_date",
+    "drs_uri",
     "file_id",
     "format",
-    "grouping_label",
     "md5_checksum",
     "size",
     "timestamp",
     "update_date"
   ],
-  "title": "file_internally_registered",
+  "title": "drs_object_registered",
   "type": "object"
 }
