+++ ghga_event_schemas/json_schemas/file_internally_registered.json
@@ -1,8 +1,8 @@
 {
-  "$id": "https://raw.githubusercontent.com/ghga-de/milestones-docs/main/milestone_1/api_definitions/async_topics/non_staged_file_requested.json",
+  "$id": "https://raw.githubusercontent.com/ghga-de/milestones-docs/main/milestone_1/api_definitions/async_topics/file_internally_registered.json",
   "$schema": "https://json-schema.org/draft/2020-12/schema",
   "additionalProperties": false,
-  "description": "This event type is triggered when a user requested to download a file from the DRS3 service that is not yet staged in the Outbox storage.",
+  "description": "This event type is triggered when a new file is registered to the Internal File Registry service and is available in the Permanent Storage.",
   "properties": {
     "creation_date": {
       "description": "Timestamp (in ISO 8601 format) when the entity was created.",
@@ -17,6 +17,10 @@
       "description": "The format of the file: BAM, SAM, CRAM, BAI, etc.",
       "type": "string"
     },
+    "grouping_label": {
+      "description": "This is a label that might be use to group multiple files together. E.g. the study id.",
+      "type": "string"
+    },
     "md5_checksum": {
       "description": "The md5-checksum of the file content.",
       "type": "string"
@@ -26,7 +30,7 @@
       "type": "integer"
     },
     "timestamp": {
-      "description": "The time when the message was published.",
+      "description": "The time when the event was published.",
       "format": "date-time",
       "type": "string"
     },
@@ -40,11 +44,12 @@
     "creation_date",
     "file_id",
     "format",
+    "grouping_label",
     "md5_checksum",
     "size",
     "timestamp",
     "update_date"
   ],
-  "title": "non_staged_file_requested",
+  "title": "file_internally_registered",
   "type": "object"
 }
