+++ ghga_event_schemas/json_schemas/file_staged_for_download.json
@@ -30,7 +30,7 @@
       "type": "integer"
     },
     "timestamp": {
-      "description": "The time when the message was published.",
+      "description": "The time when the event was published.",
       "format": "date-time",
       "type": "string"
     },
