+++ setup.cfg
@@ -15,10 +15,10 @@
 
 [metadata]
 # Please adapt to package name:
-name = ghga_message_schemas
-version = attr: ghga_message_schemas.__version__
-description = GHGA Message Schemas: A package that collects schemas used for messaging between GHGA service.
-url = https://github.com/ghga-de/ghga-message-schemas
+name = ghga_event_schemas
+version = attr: ghga_event_schemas.__version__
+description = GHGA Event Schemas: A package that collects schemas used for events exchanged between GHGA service.
+url = https://github.com/ghga-de/ghga-event-schemas
 long_description = file: README.md
 long_description_content_type = text/markdown; charset=UTF-8
 author = German Human Genome Phenome Archive (GHGA)
@@ -45,21 +45,7 @@ python_requires = >= 3.9
 
 [options.extras_require]
 dev =
-    pytest
-    pytest-cov
-    mypy
-    pylint
-    black
-    flake8
-    isort
-    bandit
-    pre-commit
-    requests
-    mkdocs
-    mkdocs-material
-    mkdocstrings
-    testcontainers
-    typer
+    ghga-service-chassis-lib[dev]==0.14.5
 all =
     %(dev)s
